+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/FoundryProperties.kt
@@ -21,6 +21,7 @@ import foundry.gradle.anvil.AnvilMode
 import foundry.gradle.artifacts.FoundryArtifact
 import foundry.gradle.properties.PropertyResolver
 import foundry.gradle.properties.getOrCreateExtra
+import foundry.gradle.properties.mapToInt
 import foundry.gradle.properties.sneakyNull
 import java.io.File
 import java.util.Locale
@@ -797,6 +798,14 @@ internal constructor(
           it
         }
       }
+
+  /** Defines JVM target to use in Kotlin and Java compilations. */
+  public val jvmTarget: Provider<Int>
+    get() =
+      resolver
+        .optionalStringProvider("foundry.jvm.target")
+        .mapToInt()
+        .orElse(resolver.gradleProviderOf(versions.jvmTarget))
 
   /** Flag to disable JVM vendor setting locally. */
   public val jvmVendorOptOut: Boolean