+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/topography/ModuleTopographyTask.kt
@@ -15,13 +15,11 @@
  */
 package foundry.gradle.topography
 
-import com.github.ajalt.mordant.markdown.Markdown
-import com.github.ajalt.mordant.rendering.AnsiLevel
-import com.github.ajalt.mordant.terminal.Terminal
 import foundry.cli.walkEachFile
 import foundry.common.json.JsonTools
 import foundry.gradle.FoundryExtension
 import foundry.gradle.FoundryProperties
+import foundry.gradle.FoundryShared
 import foundry.gradle.artifacts.FoundryArtifact
 import foundry.gradle.artifacts.Publisher
 import foundry.gradle.artifacts.Resolver
@@ -36,17 +34,23 @@ import foundry.gradle.tasks.dependsOnSourceGeneratingTasks
 import foundry.gradle.tasks.publish
 import foundry.gradle.util.toJson
 import java.nio.file.Path
+import javax.inject.Inject
 import kotlin.io.path.ExperimentalPathApi
 import kotlin.io.path.extension
 import kotlin.io.path.readText
+import kotlin.io.path.relativeTo
 import kotlin.io.path.useLines
 import kotlin.io.path.writeText
 import kotlin.jvm.optionals.getOrNull
 import org.gradle.api.DefaultTask
+import org.gradle.api.GradleException
 import org.gradle.api.Project
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.internal.plugins.PluginRegistry
+import org.gradle.api.problems.ProblemId
+import org.gradle.api.problems.Problems
+import org.gradle.api.problems.Severity
 import org.gradle.api.provider.MapProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
@@ -98,7 +102,14 @@ internal object ModuleTopographyTasks {
           DefaultFeatures.CircuitInject,
           foundryExtension.featuresHandler.circuitHandler.codegen,
         )
-        features.put(DefaultFeatures.Dagger, foundryExtension.featuresHandler.daggerHandler.enabled)
+        features.put(
+          DefaultFeatures.Dagger,
+          foundryExtension.featuresHandler.daggerHandler.enabled.zip(
+            foundryExtension.featuresHandler.daggerHandler.runtimeOnly
+          ) { enabled, runtimeOnly ->
+            enabled && !runtimeOnly
+          },
+        )
         features.put(
           DefaultFeatures.DaggerCompiler,
           foundryExtension.featuresHandler.daggerHandler.useDaggerCompiler,
@@ -162,7 +173,7 @@ public abstract class ModuleTopographyTask : DefaultTask() {
   @get:OutputFile public abstract val topographyOutputFile: RegularFileProperty
 
   init {
-    group = "foundry"
+    group = FoundryShared.FOUNDRY_TASK_GROUP
   }
 
   @TaskAction
@@ -183,7 +194,8 @@ public abstract class ModuleTopographyTask : DefaultTask() {
 }
 
 @DisableCachingByDefault
-public abstract class ValidateModuleTopographyTask : DefaultTask() {
+public abstract class ValidateModuleTopographyTask @Inject constructor(problems: Problems) :
+  DefaultTask() {
   @get:InputFile
   @get:PathSensitive(PathSensitivity.NONE)
   @get:Optional
@@ -199,12 +211,15 @@ public abstract class ValidateModuleTopographyTask : DefaultTask() {
   public abstract val autoFix: Property<Boolean>
 
   @get:Internal public abstract val projectDirProperty: DirectoryProperty
+  @get:Internal public abstract val rootDirProperty: DirectoryProperty
 
   @get:OutputFile public abstract val modifiedBuildFile: RegularFileProperty
   @get:OutputFile public abstract val featuresToRemoveOutputFile: RegularFileProperty
 
+  private val problemReporter = problems.reporter
+
   init {
-    group = "foundry"
+    group = FoundryShared.FOUNDRY_TASK_GROUP
     @Suppress("LeakingThis")
     notCompatibleWithConfigurationCache("This task modified build files in place")
     @Suppress("LeakingThis") doNotTrackState("This task modified build files in place")
@@ -283,10 +298,8 @@ public abstract class ValidateModuleTopographyTask : DefaultTask() {
 
     val allAutoFixed = featuresToRemove.all { it.replacementPatterns.isNotEmpty() }
     if (featuresToRemove.isNotEmpty()) {
-      val message = buildString {
-        appendLine(
-          "**Validation failed! The following features appear to be unused and can be removed.**"
-        )
+      val solution = buildString {
+        appendLine("The following features appear unused and can be removed:")
         appendLine()
         var first = true
         featuresToRemove.forEach {
@@ -294,27 +307,45 @@ public abstract class ValidateModuleTopographyTask : DefaultTask() {
             first = false
           } else {
             appendLine()
-            appendLine()
           }
-          appendLine("- **${it.name}:** ${it.explanation}")
-          appendLine()
-          appendLine("  - **Advice:** ${it.advice}")
+          appendLine("    ${it.name}:")
+          appendLine("        ${it.explanation}")
+          appendLine("        Advice: ${it.advice}")
         }
         appendLine()
         appendLine("Full list written to ${featuresToRemoveOutputFile.asFile.get().absolutePath}")
       }
-      val t = Terminal(AnsiLevel.TRUECOLOR, interactive = true)
-      val md = Markdown(message)
-      t.println(md, stderr = true)
       if (shouldAutoFix) {
         if (allAutoFixed) {
           logger.lifecycle("All issues auto-fixed")
         } else {
-          throw AssertionError("Not all issues could be fixed automatically")
+          report(
+            buildFile,
+            solution,
+            GradleException("Not all issues could be fixed automatically"),
+          )
         }
       } else {
-        throw AssertionError()
+        report(buildFile, solution)
       }
+    }
+  }
+
+  private fun report(
+    buildFile: Path,
+    solution: String,
+    exception: GradleException = GradleException(),
+  ) {
+    val problemId =
+      ProblemId.create(
+        "module-topography-validation",
+        "Module validation failed!",
+        FoundryShared.PROBLEM_GROUP,
+      )
+    problemReporter.throwing(exception, problemId) {
+      fileLocation(buildFile.relativeTo(rootDirProperty.asFile.get().toPath()).toString())
+      solution(solution)
+      severity(Severity.ERROR)
     }
   }
 
@@ -373,6 +404,7 @@ public abstract class ValidateModuleTopographyTask : DefaultTask() {
           topographyJson.set(topographyTask.flatMap { it.topographyOutputFile })
           featuresConfigFile.convention(foundryProperties.topographyFeaturesConfig)
           projectDirProperty.set(project.layout.projectDirectory)
+          rootDirProperty.set(project.rootDir)
           autoFix.convention(foundryProperties.topographyAutoFix)
           featuresToRemoveOutputFile.setDisallowChanges(
             project.layout.buildDirectory.file("foundry/topography/validate/featuresToRemove.json")
@@ -397,19 +429,25 @@ public abstract class ValidateModuleTopographyTask : DefaultTask() {
   }
 }
 
-//// Usage
-// var code = "foundry { features { compose() } }"
-// code = code.replace(Regex("\\bcompose\\(\\)"), "") // remove compose()
-// code = removeEmptyBraces(code) // recursively remove empty braces
-//
-// println(code) // Should print "<nothing>"
-// TODO write tests for this
-private val EMPTY_DSL_BLOCK = "(\\w*)\\s*\\{\\s*\\}".toRegex()
+private val EMPTY_DSL_BLOCK = "(\\w+\\s*\\{\\s*\\})+".toRegex()
 
+/**
+ * Removes redundant empty braces from the string. The method iteratively replaces patterns of empty
+ * DSL blocks until no such patterns remain and trims the trailing whitespace.
+ *
+ * Example:
+ * ```
+ * val input = "block1 { } block2 { content }"
+ * val result = input.removeEmptyBraces()
+ * println(result) // Output: "block2 { content }"
+ * ```
+ *
+ * @return A new string with all occurrences of empty braces removed.
+ */
 internal fun String.removeEmptyBraces(): String {
   var result = this
   while (EMPTY_DSL_BLOCK.containsMatchIn(result)) {
-    result = EMPTY_DSL_BLOCK.replace(result, "")
+    result = EMPTY_DSL_BLOCK.replace(result, "").trimEnd()
   }
   return result
 }